Further implementation of Trident
Add coverage for SDK classes
Upload an IO performance test unit

Please note that this test unit is for educated guessing only. The IOTest does not follow proper programming or even benchmark conventions. We use this to get an idea of what IO performances *looks* like. Inferences based purely on the data without deeper analysis of how the VM implements native file handling will cause you to be oblivious to other concerns regarding performance in the application!

Skeleton IO implementation
Look at threading and design considerations
Create issues on JIRA

The amount of code that was commited in this push is not an indicator of the actual amount of effort put into it! This was written after long and hard thinking, contemplating how the API and the implementation interact on the scale of plugin development to VM internals.

Please see the following links:
https://tridentsdk.atlassian.net/projects/TRDR/issues
https://tridentsdk.atlassian.net/projects/SDKR/issues